backend/requirements.txt
flask
python-dotenv
pydantic[dotenv]
faster-whisper
torch
ffmpeg-python
werkzeug